move IETHClient to the same file as ETHClient
